Add aditional icon on the right to the intro text
Initially tried to use fontawesome for finding mirrored quote sign but upon adding it into the html layout it would have slighly different form and would not scale to the size of the initial
quote sign on the lft. Had to duplicate the quote from the left and rotate it. Please, do not blame me for heresy.